Tweaks for better Fly.io perf
This app is now running with 12 regions, and this:
`fly scale count 12 --max-per-region=1`.

This gets the best possible geo distribution on deploys, rather
than doubling up on regions. `bluegreen` deploy strategies
won't work in this mode, so the PR changes it to a `rolling`
deploy.

The app servers seem to slow down with >25 concurrent requests.