Inline `@ryanflorence/md` & `@ryanflorence/mdtut`
Markdown processing being hidden behind a private package made debugging
quite difficult and pinned us to several outdated depenendencies. It
also closed over a memory leak in Shiki that sometimes causes the app
server to crash and restart.

This commit removes the `@ryanflorence/md` and `@ryanflorence/mdtut`
packages and replaces them with inline modules adapted to deal with
these issues. It also installs up-to-date versions of unified and its
plugins. This should allow us to make changes and add features to
markdown processing much more easily in the future.

Further, to deal with the memory leak we moved the shiki process into
a worker so that leaks should no longer affect the main process. This
was inspired by similar work Kent did, as his site used similar MD
processing techniques and faces similar issues. See https://kentcdodds.com/blog/fixing-a-memory-leak-in-a-production-node-js-app