Inline `@ryanflorence/md` & `@ryanflorence/mdtut`
Markdown processing being hidden behind a private package made debugging
quite difficult and pinned us to several outdated depenendencies. It
also closed over a memory leak in Shiki that sometimes causes the app
server to crash and restart.

This commit removes the `@ryanflorence/md` and `@ryanflorence/mdtut`
packages and replaces them with inline modules adapted to deal with
these issues. It also installs up-to-date versions of unified and its
plugins. This should allow us to make changes and add features to
markdown processing much more easily in the future.