Add `react-router-dom` package
Due to the way `npm` solves peer dependencies making a flat
`node_modules`, we were able to import `react-router-dom` in the project
even when we didn't have it as dependency.

This adds the dependency to avoid possible future bugs for realying on
an undeclared dependency.

Also, bumps `react-router-dom` version from 6.21.3 to 6.22.0