Update og image
I meant to update this back when I swapped out Founders Grotesk with Inter. It seems that you have to load each font weight separately versus using the variable font weights. This primarily updates the text that was previously Founders Grotesk bold to Inter black.

I also tweaked the size and alignment of elements and added the logo as a separate element instead of being included in the background image.