Update OG image (#277)
* Update og image

I meant to update this back when I swapped out Founders Grotesk with Inter. It seems that you have to load each font weight separately versus using the variable font weights. This primarily updates the text that was previously Founders Grotesk bold to Inter black.

I also tweaked the size and alignment of elements and added the logo as a separate element instead of being included in the background image.

* Remove title element from logo

Satori renders the title element for som reason. Also, I don't think this is necessary here given that an `aria-label` is already included on the anchor tag of the website.