Misc Tailwind updates (#276)
* Update tailwindcss and prettier-plugin-tailwindcss

- Updated `tailwindcss`
- Updated `prettier-plugin-tailwindcss`
- Ran the `format` script. `prettier-plugin-tailwindcss` v6 added the functionality of removing extra spaces and duplicate classes which is a great quality of life improvement. There was one spot in the footer where the spaces removed would've caused an issue because `clsx` wasn't being used to combine classes so I updated that.

* Move tailwind classes to safelist

* Remove display font family

With the removal of Founders Grotesk and using Inter for everything, there's no need for having the additional `display` font family. Removed references to it to keep things clean.

- Removed the `display` font family from the Tailwind config.
- Removed all `font-display` classes.

* Remove jit mode config

This line is no longer needed: https://tailwindcss.com/docs/upgrade-guide#migrating-to-the-jit-engine

* Update how breakpoints are configured

This isn't technically necessary, but it's a bit more conventional to extend your theme this way then to always spread the default values. Helpful for future contributors.

* Remove `tailwind-extras` from content

* Remove `font-display` classes

* Remove `@tailwindcss/aspect-ratio` plugin

The `aspect-ratio` is widely supported today so this plugin shouldn't be necessary anymore.

* Remove unnecessary CSS rules

These are applied by default in Tailwind's preflight

* Remove reference to @tailwindcss/aspect-ratio

---------

Co-authored-by: Brooks Lybrand <brookslybrand@gmail.com>