Prevent text in the mobile popover from being truncated on an iPhone 14 Pro Max
The text in this popover menu was being truncated on my iPhone 14 Pro Max. We couldn't reproduce it on my laptop even with the screen size, but after hooking up my iPhone to my Macbook to debug, we found that adding `w-max` to the `div` fixed the issue on my iPhone and didn't cause any unintended visual regressions when retesting the popover on my laptop.